Time: 25 ms (91.79%), Space: 15.7 MB (45.68%) - LeetHub

+++ 2938-separate-black-and-white-balls/2938-separate-black-and-white-balls.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     long long minimumSteps(string s) {
         int lastIndex = s.size()-1;
-        int swaps = 0;
+        long long swaps = 0;
 
         for (int i = s.size()-1; i >= 0; i--) {
             if (s[i] == '1') {
